fish-foreign-env: git-20170324 -> git-20200209
"`^` as a redirection deprecated and will be removed in the future."
(see the changelog, under the 3.0b1 release)

The latest fish beta release (3.1b1 as of time of writing) errors when
encountering `^&1` (though the fact it is now an error has yet to be
documented by them). The plugin was updated last year to account for
this change, and with the "imminent" release of fish-shell v3.1, this
should be fixed.