aws-sam-cli: add argument to conditionally enable telemetry
If someone really wants to opt into telemetry, they can do so by setting
`enableTelemetry` to `true` (the default is `false`), in which case the wrapper
that sets `SAM_CLI_TELEMETRY` to `0` will not be created.

Note that this actually allows a user to optionally disable telemetry from the
command line or the (poorly documented) configuration in
`~/.aws-sam/metadata.json`. The downside is telemetry will be enabled at least
on the first run, causing a unique installation ID to be saved in the
configuration file.