synergy: 1.8.8 -> 1.11.0 (#80138)
This gets rid of a few patches that aren't needed any more, adds one
to make the tests work again and updates the MacOS patch. It also
introduces two builds - one with the Qt application and one without.

The patch to get the tests working will be submitted upstream and
hopefully not be needed for future releases.