python3Packages: attempt to fix build, mark as broken
Earlier spyder was updated to 4.0.0. This build fails. Adding new
dependencies, it still fails, because it needs a very specific version
of jedi. This is problematic, we cannot have these tight constraints in
the Python packages set.

We could revert to 3.x. It builds, however, I expect it to have runtime
issues due to jedi as well.

https://github.com/spyder-ide/spyder/issues/11047