nixos/release-small: migrate constituents list to strings
As of a few days ago hydra has a different GC configuration that leads
to eval failures when the tested job has a direct dependency on all the
tests. This is an attempt at migrating the old list of jobs to the new
format while keeping the logic and the tests the same.