nixos/version: obfuscate stateVersion to discourage modification
Many users get the impression of having to update stateVersion when
upgrading their channel, and cause subtle or obvious breakage.

This happened repeatedly, despite nixos-generate-config adding
5 lines of comment around the definition.

Because the value of stateVersion is a NixOS version, e.g. "19.09",
people were tricked into believing this option sets the state ("system
version") of their system, when instead it sets the system version
of the state to stay compatible with.

This commit changes nothing about stateVersion itself, to ensure
backwards compatibility with uses of stateVersion inside and outside
of nixpkgs.

Instead, it now sets stateVersion to an expression that doesn't look
like a NixOS version, and should not invoke the notion of needing to be
updated.

Example configuration snippet:
system.stateVersion = lib.nixos.decodeStateVersion 2989;

Some effort is made to prevent decoding to invalid versions, but
those can still be set.

This is only for newly generated configurations, and the obfuscation
can be bypassed by just setting stateVersion to e.g. "20.03" directly,
without encoding/decoding.