noweb: fix installation, use placeholders
The problem was that nix passes lists as space-separated strings not as
arrays of strings, so `"${foo[@]}"` doesn't work as intended because
it's not an array. Instead we pass it in a bash array.

Also, using builtins.placeholder instead of passing "$(out)" to bash, as
that's not what we want to do (the `$(...)` is the process expansion in
bash)