buildRustCrate: fix erroneously hard-coded sourcePath
When using the new cargo fetcher dependencies are validated against the
lockfile in the source project. Unfortunately the code that does the
validation assumes that the source of the package is in the `source`
directory, which it doesn't have to be. This breaks any package that
keeps its source in another directory.

The fix is to use `$sourceRoot` instead of `source`.