symbiyosys: fix calls to external programs
48085826f broke symbiyosys when it needed to call `yosys-abc`: when
`ABCEXTERNAL` is set in the Makefile, then `yosys-abc` is not built.

But in general `sby` was just calling programs out of the ambient
environment. Fix that for most programs it can invoke: it now has a
direct dependency on boolector, aiger, abc, yosys, etc.

This also does some other minor clean up.

Signed-off-by: Austin Seipp <aseipp@pobox.com>