git-annex: apply patch to fix the build with ghc-8.8.x
Unfortunately, we cannot compile git-annex with S3 support in an
LTS-15.x environment, because the 'aws' library hasn't updated to
version 3.x of the 'network' library [1]. I tried whether it's
possible to build git-annex with an older version of network, but
the amount of overrides we'd have to configure to accomplish that
got out of hand quickly. So I disabled aws support [2]. If you
need S3 support in git-annex, please help upstream to update
'aws' so that it builds with recent versions of 'network'.

[1] https://github.com/aristidb/aws/issues/264
[2] https://github.com/NixOS/cabal2nix/commit/1d0459f40e0fa7483fc5c482dcbca1e1db4c40db