MissingH: use a version appropriate to our compiler
If you actually look at the changelog for 1.4.2.1, you'll see that it
mentions that it was only intended for ghc-7.0.4, which is why it has
a dependency on an inconveniently old version of `unix`, which
is *also* hidden behind a flag, so that we can't jailbreak it (per
https://github.com/peti/jailbreak-cabal/issues/15).

So the right thing to do would appear to me to be to override our
default version to 1.4.2.0, until such time as the maintainer makes a
newer release.

`ConfigFile` was collateral damage, so unmark it broken, as well as
`MissingH` itself.