httpie: add manpage and html docs
httpie has quite extensive documentation in form of its rst readme.
With docutils it’s quite easy to convert to html.

We can also convert it to a manpage, which makes the documentation
accessible from the command line (even though it’s not quite standard
format as far as manpages go; however a lot better than nothing).