gsasl: don't check on linux either
Recently something has caused the linux version to start segfaulting
in its test suite:
```
Header version 1.8.0 library version 1.8.0
/nix/store/8gy2bmpz3qawxr3g8hqhfgkf32wb0wfl-bash-4.4-p23/bin/bash: line 5: 28803 Segmentation fault      (core dumped) SHISHI_KEYS=./gssapi.key SHISHI_TICKETS=./gssapi.tkt SHISHI_CONFIG=./shishi.conf SHISHI_HOME=. SHISHI_USER=ignore-this-warning THREADSAFETY_FILES=`ls ../lib/*/*.c | /nix/store/8z2qcxbl9kfqhqqjm04lzlvvlwli4bw7-gnugrep-3.4/bin/grep -v -e lib/gl/vasnprintf.c -e lib/gl/getdelim.c` MD5FILE=./cram-md5.pwd EGREP="/nix/store/8z2qcxbl9kfqhqqjm04lzlvvlwli4bw7-gnugrep-3.4/bin/grep -E" GNUGSS=`if grep 'HAVE_LIBGSS 1' ../lib/config.h > /dev/null; then echo yes; else echo no; fi` ${dir}$tst
FAIL: old-simple
```
This library has been marked as don't-check on darwin for more than a
year---unfortunately that commit message didn't mention *why*, so it's
hard to say if linux has started seeing the same issue or a different
one.

Regardless, this is just a stopgap; the real solution is probably to
upgrade at least to 1.8.1 (a new stable release made last year), or
perhaps even to the 1.9 betas---but that seems like something that
should be the decision of the maintainer, @shlevy.