ebtables: restore "legacy" commands
ebtables 2.0.11 renamed the ebtables commands
from "ebtables*" to "ebtables-legacy-*".

Of course this breaks legacy packages and scripts
that depends on the ebtables commands.

The idea behind this upstream change appears to be that
ebtables-nft replaces ebtables and distributions should
rename either the ebtables-legacy or ebtables-nft commands
to provide the ebtables commands.

For nix a better fit is for packages to specify either the
ebtables or the ebtables-nft package, while both packages
provide the same commands.

This patch restores the ebtables package so it functions again.