staccato: remove broken and unmaintained package
This package was added in 2017, then never updated, and no longer builds.
While upstream does have a newer release at version 0.1.9, this doesn't build
with the latest rustc either, and was last released in 2018:
https://github.com/tshlabs/staccato/releases

Since the upstream package just has 1 contributor, and the package in Nix has no
maintainer and is not building or updated, let's just delete it.