sage: fix test suite with docutils 0.15
The better way to fix this would be to backport the upstream sphinx
patch:

https://github.com/sphinx-doc/sphinx/commit/faedcc48ccb942b9a7b758b699b30f0d026c0771

Unfortunately it doesn't apply cleanly and isn't worth the effort
of backporting. Let's hope we can switch to python3 sage and the recent
sphinx version that comes with it before this becomes a problem.

(cherry picked from commit 71335774051655fae820bc94c3e06c39dc778b99)