habitat: 0.30.2 -> 0.90.6
The build has been broken on master for a very long time, for want of an
upgrade. This continues the effort put forth here:
https://github.com/NixOS/nixpkgs/pull/63105

Note that while there are newer versions than this, upgrading to those requires
getting some Cargo protobuf dependencies to compile. Since this fixes the
failing build, it seems like a good start.

Upstream does not appear to have any Darin builds in their CI, and we do not
have a Darwin maintainer available, so this disables it as well.