mysql-workbench: 8.0.15 → 8.0.19
* https://dev.mysql.com/doc/relnotes/workbench/en/news-8-0-16.html
* https://dev.mysql.com/doc/relnotes/workbench/en/news-8-0-17.html
* https://dev.mysql.com/doc/relnotes/workbench/en/news-8-0-18.html
* https://dev.mysql.com/doc/relnotes/workbench/en/news-8-0-19.html

The release notes contain the following:

Important Note: MySQL Workbench 8.0.19 is unable to open a new connection to MySQL Server from the home screen if the server is not started and you cannot start a server by using the Workbench Administration feature. Administrative and SQL editing tasks require an online server for the duration of this issue. Feature tasks that you performed with an offline server in previous releases now return an error message indicating that the server is unreachable.

but it is clearly better than having it broken.

(cherry picked from commit 070b49ed00b090b444f77d50d07acc34749ebe7c)