nixos/home-assistant: fix bug #82076, change default config behaviour
If the `config` option is not set, the `port` option has no effect. This
is caused by the condition `cfg.config != null` not passing, resulting
in the default `configuration.yaml` file not being overwritten.

If `time.timeZone` isn't set, it also appears to cause home-assistant to
error out.

Another issue is that the `defaultConfig` doesn't currently set `frontend`
which means in the minimal setup with just `port` set, we still won't be able
to reach the frontend.