nettools: 1.60_p20170221182432 -> 1.60_p20180626073013
Because _p2017... cannot be fetched: 404.

The change is minimal: Adjusting whether usage messages are printed to
stdout or stderr.

It's hard to find stable source links for nettools and we resort to
fetching patched sources out of other distros' tarball caches because
this package is deprecated:
  https://lwn.net/Articles/710533/
  https://lwn.net/Articles/710535/
  https://wiki.linuxfoundation.org/networking/iproute2

The advice is to switch to iproute2, but NixOS activation scripts use
`domainname` from nettools, for which neither iproute2 nor systemd
offers a replacement.  I inquired about this on iproute2's mailing list:
  https://lore.kernel.org/netdev/CAPwpnyTDpkX2hxiqYLxTuMM38cq+whPSC0yoee-YPLEAwfvqpQ@mail.gmail.com/