rust_1_38, rust_1_42: drop patch version from attr
I find it extremely unlikely we'll ever package two different patch
versions of the same minor Rust version.  New patch versions should
generally be compatible, so we shouldn't give people the expectation
they'll be able to pin to one.  And by including the patch version in
the attribute name, we'd have to change the attribute every time a
patch version was released, which would unnecessarily create diff
noise and maintenance headaches for what should be a seamless upgrade.