rdedup: fix broken build
In the Rust platform builder, the attribute `cargoPatches` must be used if we
want to consistently apply a patch to the `Cargo.lock` in the vendored
`cargoDeps` and to the `Cargo.lock` in the `src` attribute.

This fixes the build by applying that patch consistently, so the coherency
checker in the newer Rust cargo builder passes.