kicad: don't build versions with 3d on Hydra
The hydraPlatforms have to be set on the kicad package itself, that can be
checked using:

  echo ":p { inherit kicad kicad-small kicad-unstable; }" | nix repl ./pkgs/top-level/release.nix

This commit disables build of all kicad variants that require downloading
packages3d, which currently fail on hydra with the "Output limit exceeded"
status. This leaves Hydra with only building the kicad-small, which will allow
us to cache the build of kicad-base as well as all libraries except of
packages3d.

(cherry picked from commit ebe5f107942c82386c64e170c001872a85e08ec0)