python.pkgs.httpretty: disable flaky test suite
There are some tests that fail occasionally because the current date
(with second precision) occurs in the output. This is fixed upstream.
I've made this conditional since python updates are often done in an
automated fashion and tests should be re-enabled after the next update.
This avoids the need for human intervention (except for a clean up at
some point in the future).