pydf: install man page and default config file
This makes the pydf derivation install the package's man page and default config
file, so that `man pydf` works and the user has a default config file to work
from.  This default config file is placed in `share/pydf/` and isn't read by
default; pydf still looks in `/etc` and `$HOME` instead.  (The file is identical
to the defaults in the executable itself except for additional colours applied
to the `full_fs_colour` setting.)